+++ src/app/src/reducers/ClaimFacilityDashboardReducer.jsx
@@ -6,6 +6,25 @@ import {
     failFetchFacilityClaims,
     completeFetchFacilityClaims,
     clearFacilityClaims,
+    startFetchSingleFacilityClaim,
+    failFetchSingleFacilityClaim,
+    completeFetchSingleFacilityClaim,
+    clearSingleFacilityClaim,
+    startApproveFacilityClaim,
+    startDenyFacilityClaim,
+    startRevokeFacilityClaim,
+    failApproveFacilityClaim,
+    failDenyFacilityClaim,
+    failRevokeFacilityClaim,
+    completeApproveFacilityClaim,
+    completeDenyFacilityClaim,
+    completeRevokeFacilityClaim,
+    resetFacilityClaimControls,
+    startAddNewFacilityClaimReviewNote,
+    failAddNewFacilityClaimReviewNote,
+    completeAddNewFacilityClaimReviewNote,
+    clearFacilityClaimReviewNote,
+    updateFacilityClaimReviewNote,
 } from '../actions/claimFacilityDashboard';
 
 const initialState = Object.freeze({
@@ -19,6 +38,38 @@ const initialState = Object.freeze({
         fetching: false,
         error: null,
     }),
+    statusControls: Object.freeze({
+        fetching: false,
+        error: null,
+    }),
+    note: Object.freeze({
+        note: '',
+        fetching: false,
+        error: null,
+    }),
+});
+
+const handleStartChangeStatus = state => update(state, {
+    statusControls: {
+        fetching: { $set: true },
+        error: { $set: initialState.statusControls.error },
+    },
+});
+
+const handleFailChangeStatus = (state, error) => update(state, {
+    statusControls: {
+        fetching: { $set: initialState.statusControls.fetching },
+        error: { $set: error },
+    },
+});
+
+const handleCompleteChangeStatus = (state, data) => update(state, {
+    statusControls: {
+        $set: initialState.statusControls,
+    },
+    detail: {
+        data: { $set: data },
+    },
 });
 
 export default createReducer({
@@ -43,5 +94,72 @@ export default createReducer({
     }),
     [clearFacilityClaims]: state => update(state, {
         list: { $set: initialState.list },
+    }),
+    [startFetchSingleFacilityClaim]: state => update(state, {
+        detail: {
+            fetching: { $set: true },
+            error: { $set: initialState.detail.error },
+        },
+    }),
+    [failFetchSingleFacilityClaim]: (state, error) => update(state, {
+        detail: {
+            fetching: { $set: initialState.detail.fetching },
+            error: { $set: error },
+        },
+    }),
+    [completeFetchSingleFacilityClaim]: (state, data) => update(state, {
+        detail: {
+            fetching: { $set: initialState.detail.fetching },
+            error: { $set: initialState.detail.error },
+            data: { $set: data },
+        },
+    }),
+    [clearSingleFacilityClaim]: state => update(state, {
+        detail: { $set: initialState.detail },
+    }),
+    [startApproveFacilityClaim]: handleStartChangeStatus,
+    [startDenyFacilityClaim]: handleStartChangeStatus,
+    [startRevokeFacilityClaim]: handleStartChangeStatus,
+    [failApproveFacilityClaim]: handleFailChangeStatus,
+    [failDenyFacilityClaim]: handleFailChangeStatus,
+    [failRevokeFacilityClaim]: handleFailChangeStatus,
+    [completeApproveFacilityClaim]: handleCompleteChangeStatus,
+    [completeDenyFacilityClaim]: handleCompleteChangeStatus,
+    [completeRevokeFacilityClaim]: handleCompleteChangeStatus,
+    [resetFacilityClaimControls]: state => update(state, {
+        statusControls: {
+            $set: initialState.statusControls,
+        },
+    }),
+    [startAddNewFacilityClaimReviewNote]: state => update(state, {
+        note: {
+            fetching: { $set: true },
+            error: { $set: initialState.note.error },
+        },
+    }),
+    [failAddNewFacilityClaimReviewNote]: (state, error) => update(state, {
+        note: {
+            fetching: { $set: false },
+            error: { $set: error },
+        },
+    }),
+    [completeAddNewFacilityClaimReviewNote]: (state, data) => update(state, {
+        note: {
+            $set: initialState.note,
+        },
+        detail: {
+            data: { $set: data },
+        },
+    }),
+    [clearFacilityClaimReviewNote]: state => update(state, {
+        note: {
+            $set: initialState.note,
+        },
+    }),
+    [updateFacilityClaimReviewNote]: (state, note) => update(state, {
+        note: {
+            note: { $set: note },
+            error: { $set: initialState.note.error },
+        },
     }),
 }, initialState);