+++ src/django/api/exceptions.py
@@ -1,0 +1,7 @@
+from rest_framework.exceptions import APIException
+
+
+class BadRequestException(APIException):
+    status_code = 400
+    default_detail = 'Bad request'
+    default_code = 'bad_request'